requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
Django~=4.2
django-bootstrap4==23.1
flake8==6.0.0
gunicorn==20.1.0
hexlet-code==0.1.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
soupsieve==2.4.1
sqlparse==0.4.4

django-filter~=23.2